% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpimporterr.R
\name{gdpimporterr}
\alias{gdpimporterr}
\title{Download the zipped file, unzip, rename the unzipped files, and
outputs a data frame along with the title from meta data.}
\usage{
gdpimporterr(url, filename = NULL, filetype = "csv")
}
\arguments{
\item{url}{A character vector with one element: URL to
the zip file (ends with .zip)}

\item{filename}{A character vector with one element: specify
the name of csv data (not the MetaData). If \code{NULL},
'open_canada_data.csv' will be the filename.
This argument is not useful when \code{filetype} is set to \code{'all'}}

\item{filetype}{A character vector with one element: the
types of files that will be extracted. If \code{'csv'}, only csv
files are extracted'. If \code{'all'}, files of all types are extracted}
}
\value{
A list whose first element is a data frame and
second element is a character with one element
}
\description{
Download the zipped file, unzip, rename the unzipped files, and
outputs a data frame along with the title from meta data.
}
\examples{
gdpimporterr("https://www150.statcan.gc.ca/n1/tbl/csv/36100400-eng.zip")
}
