% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpplotterr.R
\name{gdpplotterr}
\alias{gdpplotterr}
\title{gpdplotterr
Author: Aldo Saltao Barros
Plot a line chart of periods and the selected feature.}
\usage{
gdpplotterr(cleaned_Dataframe)
}
\arguments{
\item{cleaned_Dataframe}{This is the name of the dataset that you want to draw a chart. It is an output from the function gdpcleanerr}
}
\value{
line chart of GDP by the Region VS periods.
}
\description{
gpdplotterr
Author: Aldo Saltao Barros
Plot a line chart of periods and the selected feature.
}
\examples{
data <- tibble::tribble(~REF_DATE,~GEO,~DGUID,
~`North.American.Industry.Classification.System.(NAICS)`,
~UOM, ~UOM_ID,~SCALAR_FACTOR,~SCALAR_ID,~VECTOR, ~COORDINATE,
~VALUE,~STATUS,~SYMBOL,~TERMINATED,~DECIMALS,1997L,"Canada",
"2016A000011124","Total industries", "Dollars",81L,"millions",
6L, "v41894235",1.477, 816756.1, NA,NA, NA,1L,1997L,"Canada",
"2016A000011124", "Crop and animal production",
"Dollars",81L,"millions",6L,"v41700857", 1.1,12755.9,NA,NA, NA,1L)
cleaned_df <- gdpcleanerr(data)
gdpplotterr(cleaned_df)
}
